Feature #66: Exibicao de graficos de acompanhamento

+++ Ferrero.GestorDeProjetos.Web/Models/ChartJs/Datasets.cs
@@ -7,7 +7,7 @@ namespace Ferrero.GestorDeProjetos.Web.Models.ChartJs
         public string[] borderColor { get; set; }
 
         public string borderWidth { get; set; }
-        public double[] data { get; set; }
+        public decimal[] data { get; set; }
         public string fill { get; set; }
     }
 }
